+++ examples/nextjs-comments-website-overlay/src/app/page.tsx
@@ -1,7 +1,7 @@
+import { Room } from "./Room";
 import { PageHome } from "@/components/site/PageHome";
-import { Room } from "@/components/Room";
-import { Comments } from "@/components/comments/Comments";
 import { Cursors } from "@/components/cursors/Cursors";
+import { Comments } from "@/components/comments/Comments";
 
 export default async function Home() {
   return (
